+++ client/components/_Modal/LoginModal/index.tsx
@@ -1,1 +1,1 @@
-export { default } from "./loginModal";
+export { default } from "./loginModalContainer";
