+++ client/pages/_document.tsx
@@ -15,6 +15,7 @@ class MyDocument extends Document {
       <Html>
         <Head />
         <body data-theme={theme}>
+          <div id="toast-root" />
           <Main />
           <NextScript />
           <div id="portal" />
